+++ extensions/badges/src/types.ts
@@ -25,7 +25,56 @@ export type IconData = {
   aliases?: Aliases;
 };
 
+export type StaticBadge = {
+  label?: string;
+  message?: string;
+  color?: string;
+  labelColor?: string;
+  style?: string;
+  logo?: string;
+  logoColor?: string;
+  logoSize?: string;
+};
+
+export type DynamicBadge = {
+  $type?: string;
+  url?: string;
+  query?: string;
+  prefix?: string;
+  suffix?: string;
+  style?: string;
+  logo?: string;
+  logoColor?: string;
+  logoSize?: string;
+  label?: string;
+  labelColor?: string;
+  color?: string;
+};
+
+export type EndpointBadge = {
+  url?: string;
+  style?: string;
+  logo?: string;
+  logoColor?: string;
+  logoSize?: string;
+  label?: string;
+  labelColor?: string;
+  color?: string;
+};
+
+export type SimpleIcons = {
+  $icon?: IconData;
+};
+
+export type Badge = SimpleIcons & StaticBadge & DynamicBadge & EndpointBadge;
+
 export type LaunchFromSimpleIconsContext = {
   launchFromExtensionName: string;
   icon: IconData;
+};
+
+export type LaunchFromColorPickerContext = {
+  launchFromExtensionName: string;
+  field: keyof Badge;
+  hex: string;
 };