+++ extensions/date-converter/src/index.tsx
@@ -2,6 +2,11 @@ import { Clipboard, ActionPanel, List, Action, getPreferenceValues } from "@rayc
 import * as chrono from "chrono-node";
 import { useMemo, useState } from "react";
 import "@total-typescript/ts-reset";
+import TimeAgo from "javascript-time-ago";
+import en from "javascript-time-ago/locale/en";
+
+TimeAgo.addDefaultLocale(en);
+const timeAgo = new TimeAgo("en-US");
 
 interface LabeledDate {
   label?: string;
@@ -19,11 +24,13 @@ interface DateFormatter {
 interface Preferences {
   defaultFormat: string;
   copyAction: "copy" | "paste" | "both";
+  hour24: boolean;
 }
 
-const humanFormatter = new Intl.DateTimeFormat([], {
+const humanFormatter = new Intl.DateTimeFormat(undefined, {
   dateStyle: "full",
-  timeStyle: "short",
+  timeStyle: "medium",
+  hour12: !getPreferenceValues<Preferences>().hour24,
 });
 
 const DATE_FORMATS: DateFormatter[] = [
@@ -183,7 +190,7 @@ export default function Command() {
         <List.Item
           key={date.toISOString()}
           title={humanFormatter.format(date)}
-          subtitle={label}
+          subtitle={`${label} - ${timeAgo.format(date)}`}
           actions={
             <ActionPanel>
               {getSortedFormats({ human }).map(({ id, title, format }) => (