+++ extensions/mute-microphone/src/toggle-mute.tsx
@@ -1,13 +1,12 @@
 import { LaunchType, closeMainWindow, launchCommand } from "@raycast/api";
-import { get, toggleSystemAudioInputLevel } from "./utils";
+import { toggleSystemAudioInputLevel } from "./utils";
 
 export default async function toggleMute() {
-  const currentAudioInputLevel = await get();
   await closeMainWindow();
-  await toggleSystemAudioInputLevel(Number(currentAudioInputLevel));
+  await toggleSystemAudioInputLevel();
 
   try {
-    await launchCommand({ name: "mute-menu-bar", type: LaunchType.Background });
+    await launchCommand({ name: "mute-menu-bar", type: LaunchType.UserInitiated });
   } catch {
     console.log("mute-menu-bar command is not active");
   }