+++ extensions/1password/CHANGELOG.md
@@ -1,5 +1,17 @@
 # 1Password Changelog
 
+## [Fix] - 2024-05-23
+
+-- Fixes the infinite loading when trying to switch accounts.
+
+## [Fix] - 2024-05-21
+
+-- Fixes account_uuid is undefined
+
+## [Fix] - 2024-05-21
+
+-- Fixes account selection
+
 ## [Enhancements] - 2024-05-10
 
 -- Fixes double auth prompts.