+++ extensions/groq/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Groq Changelog
 
+## [Update] - 2024-05-18
+
+- Added new AI chat command `AI Chat`.
+- Improved system prompt for `Fix Grammar and Spelling` and `Improve Text` commands.
+
 ## [Update] - 2024-04-25
 
 - Question of command `Ask AI` now displayed at the top of the response