+++ extensions/linear/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Linear Changelog
 
+## [Improvements] - {PR_MERGE_DATE}
+
+- Added preference to hide redundant (Done, Canceled, Duplicate) issues across all searches
+
+## [Bug Fixes] - 2024-05-25
+
+- Fixed milestones to be updated with "Create Issue" command
+
 ## [Improvements] - 2024-05-14
 
 - Notifications now support Projects (Updates, comments, reactions, added as member), Document Mentions and Bots (GitHub and GitLab)